Update code in RoomController, RoomService, Add unit tests for RoomController

+++ src/main/java/ch/uzh/ifi/hase/soprafs23/rest/dto/RoomGetDTO.java
@@ -7,7 +7,7 @@ public class RoomGetDTO {
     private int roomId;
     private ArrayList<Long> userIds = new ArrayList<>();
 
-    public long getRoomId() {
+    public int getRoomId() {
         return roomId;
     }
